+++ codemods/transforms/__tests__/linkToHref.test.ts
@@ -1,0 +1,7 @@
+import {defineTest} from 'jscodeshift/src/testUtils';
+
+const testName = 'linkToHref';
+
+defineTest(__dirname, testName, null, testName, {
+    parser: 'tsx',
+});