+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [3.7.1](https://github.com/gravity-ui/navigation/compare/v3.7.0...v3.7.1) (2025-09-29)
+
+
+### Bug Fixes
+
+* clean up resize observer ([#461](https://github.com/gravity-ui/navigation/issues/461)) ([72ff021](https://github.com/gravity-ui/navigation/commit/72ff0214644e6a947ff3e91a67f9c316d7cb74b7))
+
+## [3.7.0](https://github.com/gravity-ui/navigation/compare/v3.6.4...v3.7.0) (2025-09-11)
+
+
+### Features
+
+* **AsideHeader:** add possibility to provide className to the AsideHeaderItem ([#464](https://github.com/gravity-ui/navigation/issues/464)) ([d52b43d](https://github.com/gravity-ui/navigation/commit/d52b43d453ec147e7c1f1e156cfa6ea51125454d))
+
 ## [3.6.4](https://github.com/gravity-ui/navigation/compare/v3.6.3...v3.6.4) (2025-08-11)